Tests: Optionally extend RunCMake.ExternalProject download server timeout
Some of the `RunCMake.ExternalProject` test cases use a download server
implemented in Python.  Normally we wait up to 30 seconds for it to
start.  However, on some machines running many tests concurrently, the
download server may take longer than that to start.  Add an undocumented
cache entry to use on those machines to extend the timeout.