Utilities/Git: add a file for refs to ignore when blaming
This file may be passed to `git blame` through either the
`--ignore-revs-file` argument or the `blame.ignoreRevsFile`
configuration to ignore certain commits when performing blames.

Note that if Git is unable to find an older commit for a given line,
commits in this file may still be given as the source of a line.

The file uses features (namely empty lines and comments) which were
added in Git 2.20, released Dec 2018.