GHS: Set LANGUAGE type for source files
Use gbuild project syntax for setting the language type for the file.
gbuild will pick the correct compiler or tool to use.

NOTE: Language names come from gbuild config files.  Therefore to
compile a file with the C++ compiler use "C++" not "CXX".

The previous implementation was passing the C compiler a flag to compile as C++
if LANGUAGE was set to "CXX".