LCC: get rid of excess -Wunused-but-set-variable on release builds
LCC, as of 1.26, has a buggy EDG frontend that emits
-Wunused-but-set-variable warning even if there's something
like "(void)var" when frontend optimizations are enabled (>=-O1).
To get rid of this warning, we disable it in CMake's root
CMakeLists.txt for LCC >= 1.26. Later, when this bug will be fixed
in LCC, this version check will have its top version limit.